Removed sprite RenderQueue and moved FrameBuffer to rendering instead of rendering/util
